+++ data/examples/other/disabling/normalize-magic-comments-out.hs
@@ -1,0 +1,8 @@
+{- ORMOLU_DISABLE -}
+foo=bar
+{- ORMOLU_ENABLE -}
+bar = baz
+
+{- ORMOLU_DISABLE -}
+baz=foo
+{- ORMOLU_ENABLE -}