+++ data/examples/other/disabling/normalize-magic-comments.hs
@@ -1,0 +1,7 @@
+  {-ORMOLU_DISABLE   -}
+foo=bar
+{-   ORMOLU_ENABLE-}
+bar=baz
+{-   ORMOLU_DISABLE-}
+baz=foo
+    {-ORMOLU_ENABLE     -}