+++ src/Ormolu/Processing/Postprocess.hs
@@ -1,5 +1,4 @@
 {-# LANGUAGE OverloadedStrings #-}
-{-# LANGUAGE ViewPatterns #-}
 
 -- | Postprocessing for the results of printing.
 module Ormolu.Processing.Postprocess
@@ -7,6 +6,7 @@ module Ormolu.Processing.Postprocess
   )
 where
 
+import Data.Char (isSpace)
 import Data.Text (Text)
 import qualified Data.Text as T
 import Ormolu.Processing.Common
@@ -16,16 +16,30 @@ import qualified Ormolu.Processing.Cpp as Cpp
 postprocess ::
   -- | Desired indentation level
   Int ->
+  -- | Regions where formatting was disabled.
+  DisabledRegions ->
   -- | Input to process
   Text ->
   Text
-postprocess indent =
+postprocess indent disabledRegions =
   T.unlines
-    . fmap indentLine
+    . pasteDisabledRegions
+    . fmap (indentLine indent)
     . fmap Cpp.unmaskLine
-    . filter (not . magicComment)
     . T.lines
   where
-    magicComment (T.stripStart -> x) =
-      x == startDisabling || x == endDisabling
-    indentLine x = T.replicate indent " " <> x
+    magicComment = (== disableMarker) . T.stripStart
+    indentLine n x = T.replicate n " " <> x
+    pasteDisabledRegions = paste disabledRegions id
+    paste regions'@(r : regions) outputSoFar (l : ls)
+      | magicComment l =
+        let indentationLevel = indent + T.length (T.takeWhile isSpace l)
+         in paste
+              regions
+              ( outputSoFar
+                  . (l :)
+                  . ((indentLine indentationLevel . T.pack <$> r) ++)
+              )
+              ls
+      | otherwise = paste regions' (outputSoFar . (l :)) ls
+    paste _ outputSoFar remainingLines = outputSoFar remainingLines